+++ outputs.tf
@@ -1,3 +1,8 @@
+output "data_disk_ids" {
+  description = "The list of data disk IDs attached to this Virtual Machine."
+  value       = [for attachment in azurerm_virtual_machine_data_disk_attachment.attachment : attachment.id]
+}
+
 output "network_interface_id" {
   description = "Id of the vm nic that created by this module. `null` if `var.network_interface_ids` is provided."
   value       = try(azurerm_network_interface.vm[0].id, null)
@@ -51,9 +56,4 @@ output "vm_virtual_machine_scale_set_id" {
 output "vm_zone" {
   description = "The Availability Zones in which this Virtual Machine is located."
   value       = local.virtual_machine.zone
-}
-
-output "data_disk_ids" {
-  description = "The list of data disk IDs attached to this Virtual Machine."
-  value       = [for attachment in azurerm_virtual_machine_data_disk_attachment.attachment : attachment.id]
 }