+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 **Merged pull requests:**
 
+- add VM Extension dependency on data disks attachment and output list of attached disk\(s\) \#33 [\#35](https://github.com/Azure/terraform-azurerm-virtual-machine/pull/35) ([tl-aiyor](https://github.com/tl-aiyor))
 - Remove precondition on boot\_diagnostics [\#32](https://github.com/Azure/terraform-azurerm-virtual-machine/pull/32) ([lonegunmanb](https://github.com/lonegunmanb))
 - add tracing tag toggle variables and update readme [\#28](https://github.com/Azure/terraform-azurerm-virtual-machine/pull/28) ([lonegunmanb](https://github.com/lonegunmanb))
 - Bump github.com/Azure/terraform-module-test-helper from 0.12.0 to 0.13.0 in /test [\#25](https://github.com/Azure/terraform-azurerm-virtual-machine/pull/25) ([dependabot[bot]](https://github.com/apps/dependabot))